feat: extract encryption operations to dedicated service
Extract encryption operations (~140 lines) from frontend.rs into new
encryption/service.rs module with EncryptionService. Frontend now uses
EncryptionService instances to handle parameter encryption, literal
encryption, and statement analysis for encrypted operations.

- Created encryption/service.rs with EncryptionService
- Extracted encrypt_literals, encrypt_params, to_encryptable_statement, literals_to_plaintext methods
- Updated frontend.rs to use EncryptionService::new(&encrypt, keyset_id)
- Optimized to reuse EncryptionService instances within methods
- Cleaned up unused imports (ColumnMapper, Column, literal_from_sql, EqlTerm, Plaintext)
- Removed client_id from encryption debug messages

Part of incremental frontend refactoring (Step 4/7).
