+++ packages/cipherstash-proxy/src/postgresql/encryption/mod.rs
@@ -1,0 +1,3 @@
+pub mod service;
+
+pub use service::EncryptionService;
